src/requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.45
botocore==1.19.45
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
codecov==2.1.10
colorama==0.4.4
coverage==5.3
fpl==0.6.25
idna==2.10
iniconfig==1.1.1
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.0.2
packaging==20.7
pluggy==0.13.1
PTable==0.9.2
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
pytest-aiohttp==0.3.0
pytest-cov==2.10.1
pytest-mock==3.3.1
python-dateutil==2.8.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
yarl==1.6.3
